Organized files. Changed all links
We only have one html file that is not index.html that is not in index.html. This is because Max is currently working on it and I do not want to deal with more merge conflicts then we need to. Will manually move that file over when Max is finished.